requirements.txt
adjustText==1.3.0
aiofiles==24.1.0
aiohttp==3.9.3
aiomysql==0.2.0
aiosignal==1.3.1
alembic==1.13.3
altgraph==0.17.4
amqp==5.2.0
annotated-types==0.6.0
anthropic==0.34.2
anyio==3.7.1
appdirs==1.4.4
appnope==0.1.3
asgiref==3.8.1
astroid==3.3.5
asttokens==2.2.1
asyncio==3.4.3
attrs==23.2.0
auth0-python==4.7.2
autopep8==2.3.1
av==12.3.0
babel==2.17.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
billiard==4.2.1
black==24.10.0
blinker==1.6.3
Brotli==1.1.0
build==1.2.2.post1
CacheControl==0.13.1
cachetools==5.3.1
celery==5.4.0
certifi==2023.5.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.4.1
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudscraper==1.2.71
cohere==5.11.3
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.4
configparser==7.1.0
contourpy==1.1.0
courlan==1.3.2
crayons==0.4.0
crewai==0.76.9
crewai-tools==0.13.4
cryptography==43.0.3
cssselect==1.3.0
cssutils==2.11.0
ctranslate2==4.3.1
cycler==0.11.0
dataclasses-json==0.6.7
datasets==2.20.0
dateparser==1.2.1
debugpy==1.6.7.post1
decorator==4.4.2
defusedxml==0.7.1
dependency-injector==4.46.0
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.8
distro==1.9.0
dnspython==2.3.0
docarray==0.40.0
docker==7.1.0
docstring_parser==0.16
docx2txt==0.8
durationpy==0.9
ecdsa==0.19.0
email_validator==2.2.0
embedchain==0.1.124
et-xmlfile==1.1.0
executing==1.2.0
facebook_page_scraper==5.0.6
Faker==19.6.2
fastapi==0.115.12
fastapi-cli==0.0.4
fastavro==1.9.7
faster-whisper==1.0.3
feedfinder2==0.0.4
feedparser==6.0.11
ffmpeg-python==0.2.0
filelock==3.14.0
firebase-admin==6.2.0
flake8==7.3.0
Flask==3.0.3
Flask-Cors==4.0.0
Flask-Executor==1.0.0
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.42.1
frozenlist==1.4.1
fsspec==2024.5.0
future==1.0.0
google-ads==25.1.0
google-ai-generativelanguage==0.6.15
google-api-core==2.23.0
google-api-python-client==2.175.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.26.0
google-cloud-core==2.4.1
google-cloud-firestore==2.12.0
google-cloud-resource-manager==1.13.0
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-generativeai==0.8.5
google-resumable-media==2.6.0
google-search-results==2.4.2
googleapis-common-protos==1.66.0
googlemaps==4.10.0
gptcache==0.1.44
greenlet==3.0.3
grpc-google-iam-v1==0.13.1
grpcio==1.68.0
grpcio-health-checking==1.68.0
grpcio-reflection==1.68.0
grpcio-status==1.62.3
grpcio-tools==1.62.3
gspread==5.11.3
gspread-dataframe==3.3.1
gunicorn==21.2.0
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
html2markdown==0.1.7
html2text==2024.2.26
htmldate==1.9.3
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.1
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.23.4
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
imageio==2.35.1
imageio-ffmpeg==0.4.9
importlib_metadata==8.4.0
importlib_resources==6.4.5
iniconfig==2.0.0
install==1.3.5
instructor==1.6.3
ipykernel==6.25.1
ipython==8.14.0
isort==5.13.2
itsdangerous==2.1.2
jcloud==0.3
jedi==0.19.0
jieba==0.42.1
jieba3k==0.35.1
jina-hubble-sdk==0.39.0
Jinja2==3.1.5
jiter==0.5.0
joblib==1.4.2
json_repair==0.30.0
jsonpatch==1.33
jsonpickle==3.3.0
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2023.7.1
jupyter_client==8.3.0
jupyter_core==5.3.1
jusText==3.0.2
kaitaistruct==0.10
kiwisolver==1.4.5
kombu==5.4.2
kubernetes==31.0.0
lancedb==0.15.0
langchain==0.3.7
langchain-anthropic==0.2.1
langchain-cohere==0.3.1
langchain-community==0.3.5
langchain-core==0.3.15
langchain-experimental==0.3.3
langchain-openai==0.2.1
langchain-text-splitters==0.3.0
langgraph==0.2.34
langgraph-checkpoint==2.0.0
langsmith==0.1.130
lazy_loader==0.4
litellm==1.51.3
llvmlite==0.42.0
lxml==5.2.0
lxml_html_clean==0.4.1
macholib==1.16.3
Mako==1.3.6
Markdown==3.7
markdown-it-py==3.0.0
markdownify==0.12.1
MarkupSafe==2.1.3
marshmallow==3.22.0
matplotlib==3.10.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mem0ai==0.1.27
mlx==0.21.0
mlx-whisper==0.4.1
mmh3==5.0.1
monotonic==1.6
more-itertools==10.2.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.44.0
nest-asyncio==1.5.7
networkx==3.1
newspaper3k==0.2.8
nltk==3.9.1
nodeenv==1.9.1
numba==0.59.1
numpy==2.2.6
oauthlib==3.2.2
onnxruntime==1.19.2
openai==1.61.0
openai-whisper==20240930
opencv-python==4.10.0.84
openpyxl==3.1.5
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-exporter-otlp-proto-http==1.28.2
opentelemetry-exporter-prometheus==0.49b2
opentelemetry-instrumentation==0.49b2
opentelemetry-instrumentation-aiohttp-client==0.49b2
opentelemetry-instrumentation-asgi==0.49b2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-instrumentation-grpc==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opentelemetry-util-http==0.49b2
orjson==3.10.6
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.3.0
parameterized==0.9.0
parso==0.8.3
pathspec==0.12.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
playwright==1.42.0
plotly==6.1.2
pluggy==1.5.0
portalocker==2.10.1
posthog==3.7.0
proglog==0.1.10
prometheus_client==0.21.0
prompt-toolkit==3.0.39
proto-plus==1.24.0.dev1
protobuf==4.25.5
psutil==5.9.5
ptyprocess==0.7.0
pulsar-client==3.5.0
pure-eval==0.2.2
pyactiveresource==2.2.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.14.0
pycparser==2.21
pydantic==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.18.1
pyee==11.0.1
pyflakes==3.4.0
Pygments==2.16.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
PyJWT==2.8.0
pylance==0.19.1
pylint==3.3.1
pymongo==4.4.0
PyMySQL==1.1.1
pyngrok==7.1.0
pyOpenSSL==24.3.0
pyparsing==3.0.9
pypdf==5.1.0
pypdfium2==4.22.0
pyperclip==1.8.2
PyPika==0.48.9
pyproject_hooks==1.2.0
pyright==1.1.387
pysbd==0.3.4
PySocks==1.7.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.3.0
python-multipart==0.0.6
pytrends==4.9.2
pytube==15.0.0
pytubefix==8.12.1
pytz==2024.2
pyvis==0.3.2
PyYAML==6.0.1
pyzmq==25.1.1
qdrant-client==1.12.1
readability-lxml==0.8.1
referencing==0.30.2
regex==2024.9.11
requests==2.32.4
requests-file==2.1.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==14.0.0
rpds-py==0.10.6
rsa==4.9
safetensors==0.4.3
schema==0.7.7
scikit-image==0.24.0
scikit-learn==1.7.0
scipy==1.14.1
seaborn==0.13.2
selenium==4.27.1
selenium-wire==5.1.0
# Editable install with no version control (seo_report_agent==1.0.0)
-e /Users/rose/Documents/這個月要看/cursor_projects/seo-report-agent-based/seo-report-agent
serpapi==0.1.5
sgmllib3k==1.0.0
shapely==2.0.6
shellingham==1.5.4
ShopifyAPI==12.6.0
six==1.16.0
sniffio==1.3.1
snownlp==0.12.3
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.31
stack-data==0.6.2
starlette==0.46.1
sympy==1.13.1
tabulate==0.9.0
tavily-python==0.5.0
tenacity==9.0.0
threadpoolctl==3.5.0
tifffile==2024.9.20
tiktoken==0.7.0
tinysegmenter==0.3
tk==0.1.0
tld==0.13
tldextract==5.1.3
tokenizers==0.19.1
tomli==2.0.2
tomli_w==1.1.0
tomlkit==0.13.2
torch==2.6.0
tornado==6.3.3
tqdm==4.67.1
trafilatura==2.0.0
traitlets==5.9.0
transformers==4.42.3
trio==0.25.0
trio-websocket==0.11.1
typer==0.16.0
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.3.1
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.3.0
urllib3-secure-extra==0.1.0
uv==0.4.29
uvicorn==0.29.0
uvloop==0.19.0
vine==5.1.0
watchfiles==0.22.0
wcwidth==0.2.6
webdriver-manager==3.2.2
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.0.0
wordcloud==1.9.3
wrapt==1.16.0
wsproto==1.2.0
xlrd==2.0.1
xxhash==3.4.1
yapf==0.40.2
yarl==1.9.4
youtube-transcript-api==0.6.3
zipp==3.20.2
zstandard==0.23.0
